requirements.txt
streamlit
pandas
google-auth
google-auth-oauthlib
google-auth-httplib2
google-cloud-bigquery
google-auth-oauthlib
db-dtypes
